requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
python-multipart==0.0.6
anthropic==0.5.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-magic==0.4.27
aiofiles==23.2.1
loguru==0.7.2 